Check sample quality and filter the count matrix

This function is used to reads the output from cellranger, create a seurat object, apply basic filtering, and normalise and cluster the filtered counts.

run_qc_and_filtering(
  exp_meta,
  sample_num,
  pipeline,
  minFeatures,
  pcDims = 30,
  mtPerc = 20,
  source = "raw",
  source_type = "dir",
  organism = "human",
  isVerbose = FALSE,
  vst_flavor = NULL,
  input_matrix = NULL,
  isFixed = FALSE,
  remove_doublets = FALSE,
  mtPattern = NULL,
  riboPattern = NULL,
  seurat5 = FALSE,
  bpcells_dir = NULL,
  labeling_celline = FALSE
)

Arguments

exp_meta

Metadata table (data.frame by default)

sample_num

Row number of the metadata file containing the sample information

pipeline

Perform SCT or lognorm normalisation

minFeatures

Maximum number of genes to filter out

pcDims

Number of PCs (30 by default)

source

Use the data from the raw or filtered cellranger matrix (raw by default)

source_type

dir by default

organism

Set the sample organism (human by default)

isVerbose

Print all the messages on screen (FALSE by default)

vst_flavor

sctransform version

Value

Returns

a folder, named as the sample ID, containing QC figures and tables

a seurat object with filtering and clustering applied

Examples

if (FALSE) {
run_qc_and_filtering(exp_meta,number,".","SCT", 500, 30, source = "raw", source_type = "outs", isVerbose=FALSE)
}