Plot a STEMNET object by Circular A Posteriori Projection

Plot a STEMNET object by Circular A Posteriori Projection

stemnet_plot(x, arrangement = NULL, top = 1, color = "direction")

Arguments

x

An object of class stemnet

arrangement

The order of developmental endpoints on the CAP plot. If NULL (default), the optimal arrangement is determined using a travelling salesman algorithm, see below.

top

If arrangement is NULL, show the top n solutions for arranging the developmental endpoints. Defaults to 1, i.e. show only the best solution.

color

Specify the variable color-coded on the plot. The following options exist:

One of "direction" or "degree": Highlight the degree or direction of lineage priming.

A numeric vector of length equal the number of cells included in the STEMNET plot

A boolean vector to impose a density plot on the STEMNET plot

Value

Invisibly returns a list of two elements: First, the data underlying the plot; Second, the arrangement of endpoints on the cirlce.

Details

The posterior probability matrix computed by STEMNET places each cell into a n-simplex with as many corners as there a developmental endpoints. To reduce this high-dimensional space to 2 dimensions for plotting, the endpoints are arranged on the edge of a circle and all cells are placed in between.

In brief, each endpoint is assigned with an angle \(\alpha_k\). The class probabilties \(p_{ik}\) are then transformed to cartesian coordinates by $$x_i = \sum_k p_{ik} \cos (\alpha_k) $$ and $$y_i = \sum_k p_{ik} \sin (\alpha_k) $$

To find the optimal arrangement of the developmental endpoints on the circle, lineages with common precursor stages are placed next to each other. The proximity between lineages l and k computed by $$D_{kl} = \sum_i p_{il} * p_{ik}$$

The arrangement with the highest proximity is chosen.

References

D. Jaitin, E. Kenigsberg, et al.: Massively parallel single-cell RNA-seq for marker-free decomposition of tissues into cell types. *Science* 343: 776-779